backend/requirements.txt
redis
websockets
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt
django-cors-headers
psycopg2
psycopg2-binary
gunicorn
cryptography
python-jose
djangorestframework_simplejwt
absl-py
psycopg2-binary
requests
tabulate
selectolax
python-decouple
sentry-sdk
cdx-toolkit